requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pypdf==4.2.0
pymupdf
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
Pillow
chardet==5.2.0
aiofiles==24.1.0
python-multipart==0.0.9
watchdog
customtkinter
pystray
elasticsearch
